requirements.txt
asgiref==3.3.1
Django==2.2
django-ckeditor==6.0.0
django-crispy-forms==1.11.1
django-etc==1.3.0
django-hitcount==1.3.3
django-js-asset==1.2.2
django-taggit==1.3.0
django-widget-tweaks==1.4.8
Pillow==8.1.0
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
